oh-my-zsh: added
Updated zsh configuration to not conflict with oh-my-zsh configurations,
and updated install script to install oh-my-zsh. The install script
currently uses a fork of the oh-my-zsh install script that does not
automatically open a zsh environment after installation. The opening of
said zsh environment hard blocks the rest of the install script until
the environment is exited, which is confusing and annoying. This fork
will likely eventually be merged back into oh-my-zsh main, and this
install script should be updated accordingly if and when that happens.